modified:   app/Http/Controllers/SuratKeluarController.php 	modified:   app/Http/Controllers/SuratMasukController.php

+++ app/Http/Controllers/SuratKeluarController.php
@@ -231,21 +231,16 @@ class SuratKeluarController extends Controller
 
     public function deletekeluar($id)
     {
-        $suratKeluar = SuratKeluar::findOrFail($id);
-        $suratKeluar->delete();
+        $suratKeluar = SuratMasuk::findOrFail($id);
+        if (!$suratKeluar) {
+            Alert::error('Data tidak ditemukan', 'Data dengan ID yang diberikan tidak ditemukan.');
+        } else {
+            $suratKeluar->delete();
+            Alert::success('Data Berhasil Dihapus', 'Data surat masuk telah berhasil dihapus dari database.')->toHtml();
+        }
+        return redirect()->back()->with('success', 'Data surat masuk berhasil dihapus.');
 
-        return redirect()->back()->with('success', 'Data surat keluar berhasil dihapus.');
-        // $data = SuratKeluar::find($id);
-        // if (!$data) {
-        //     Alert::error('Data tidak ditemukan', 'Data dengan ID yang diberikan tidak ditemukan.');
-        // } else {
-        //     $data->delete();
-        //     Alert::success('Data berhasil dihapus', 'Data surat keluar telah dihapus.');
-        // }
-        // return redirect('/daftar-surat-keluar');
-        // $data->delete();
-        // return redirect('/daftar-surat-keluar');
-        // return redirect()->route('daftar-surat-keluar')->with('success', 'Data Berhasil di Hapus');
+
     }
 
     public function exportpdfkeluar(){
